Updated clear upstream cache button to send GraphQL mutation

See merge request https://gitlab.com/gitlab-org/gitlab/-/merge_requests/222830

Merged-by: Rahul Chanila <rchanila@gitlab.com>
Approved-by: Rahul Chanila <rchanila@gitlab.com>

+++ ee/app/assets/javascripts/packages_and_registries/virtual_registries/components/common/upstreams/table.vue
@@ -37,6 +37,7 @@ export default {
   inject: {
     editUpstreamPathTemplate: { default: '' },
     showUpstreamPathTemplate: { default: '' },
+    deleteUpstreamCacheMutation: { default: null },
   },
   props: {
     upstreams: {
@@ -127,7 +128,17 @@ export default {
       const { id } = this.selectedUpstream;
       this.hideUpstreamClearCacheModal();
       try {
-        await deleteMavenUpstreamCache({ id: getIdFromGraphQLId(id) });
+        if (this.deleteUpstreamCacheMutation) {
+          await this.$apollo.mutate({
+            mutation: this.deleteUpstreamCacheMutation,
+            variables: {
+              id,
+            },
+          });
+        } else {
+          await deleteMavenUpstreamCache({ id: getIdFromGraphQLId(id) });
+        }
+
         this.$toast.show(s__('VirtualRegistry|Upstream cache cleared successfully.'));
       } catch (error) {
         this.$toast.show(s__('VirtualRegistry|Failed to clear upstream cache. Try again.'));
